Move endpoint.Configure struct
