+++ fcoin/endpoint/public.go
@@ -2,16 +2,6 @@ package endpoint
 
 import "fmt"
 
-type Configure struct {
-	//  The adapter that will be used to connect if none is set
-	Adapter   string
-	Endpoint  string
-	UserAgent string
-	Proxy     string
-	ApiKey    string
-	SecretKey string
-}
-
 //http://horie1024.hatenablog.com/entry/2014/08/25/012123
 func PublicServerTime(c *Configure) (ret string, err error) {
 	url := Public(c) + "server-time"
